refactor(tile): rename scss files to use .module.scss
