+++ draft-packages/tile/KaizenDraft/Tile/TileGrid.module.scss
@@ -1,6 +1,6 @@
 @import "~@kaizen/design-tokens/sass/spacing";
 @import "~@kaizen/component-library/styles/responsive";
-@import "./dimensions";
+@import "./variables";
 
 .grid {
   display: grid;
