+++ draft-packages/tile/KaizenDraft/Tile/components/GenericTile.module.scss
@@ -4,7 +4,7 @@
 @import "~@kaizen/deprecated-component-library-helpers/styles/layout";
 @import "~@kaizen/component-library/styles/animation";
 @import "~@kaizen/component-library/styles/responsive";
-@import "../dimensions";
+@import "../variables";
 
 $tilePaddingTop: $spacing-xl;
